-Se cambia en Factura el numero del documento de int a long -Se organiza metodo al obtener un documento para que reciba un long
git-svn-id: http://192.168.0.251:18080/svn/HGInetMiFacturaElectronica/Tronco@662 30a10932-8c21-4e4d-906a-15261f94da71

+++ HGInetMiFacturaElectonicaController/Registros/Ctl_Documento.cs
@@ -61,7 +61,7 @@ namespace HGInetMiFacturaElectonicaController.Registros
 
 		#region Obtener
 
-		public TblDocumentos Obtener(string numero_resolucion, int numero_documeto, int tipo_doc)
+		public TblDocumentos Obtener(string numero_resolucion, long numero_documeto, int tipo_doc)
 		{
 			try
 			{
